Added stock_loss file and loss checks
