+++ rnn/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(examm_nn generate_nn.cxx rnn_genome.cxx rnn.cxx lstm_node.cxx ugrnn_node.cxx delta_node.cxx gru_node.cxx enarc_node.cxx enas_dag_node.cxx random_dag_node.cxx mgu_node.cxx dnas_node.cxx mse.cxx rnn_node.cxx rnn_edge.cxx rnn_recurrent_edge.cxx rnn_node_interface.cxx genome_property.cxx sin_node.cxx sum_node.cxx cos_node.cxx tanh_node.cxx sigmoid_node.cxx inverse_node.cxx multiply_node.cxx sin_node_gp.cxx cos_node_gp.cxx tanh_node_gp.cxx sigmoid_node_gp.cxx inverse_node_gp.cxx multiply_node_gp.cxx sum_node_gp.cxx)
+add_library(examm_nn generate_nn.cxx rnn_genome.cxx rnn.cxx lstm_node.cxx ugrnn_node.cxx delta_node.cxx gru_node.cxx enarc_node.cxx enas_dag_node.cxx random_dag_node.cxx mgu_node.cxx dnas_node.cxx mse.cxx rnn_node.cxx rnn_edge.cxx rnn_recurrent_edge.cxx rnn_node_interface.cxx genome_property.cxx sin_node.cxx sum_node.cxx cos_node.cxx tanh_node.cxx sigmoid_node.cxx inverse_node.cxx multiply_node.cxx sin_node_gp.cxx cos_node_gp.cxx tanh_node_gp.cxx sigmoid_node_gp.cxx inverse_node_gp.cxx multiply_node_gp.cxx sum_node_gp.cxx stock_loss.cxx)
 target_link_libraries(examm_nn exact_time_series exact_weights exact_common)
